requirements.txt
Flask==2.3.2
tensorflow==2.11.0
Pillow==9.4.0
numpy==1.23.5
gunicorn==20.1.0
werkzeug==2.2.3
